requirements.txt
PyExecJs
requests
baidu-aip